Adds MSVC project, fixes for Windows, Resnet
Adds a MSVC project for compiling a static library.
Adds fixes so that things will work on Windows.
Attempted to poke around in OpenMP code to see if I could port that to Windows, but there's heavy dependence on pthread so I just butchered stuff before giving up.
Adds a fix that is from the main caffe branch so that "legacy" layers in Yahoo's OpenNSFW will load. Without this, the model doesn't load.